EPCISEvent inheritance properties
