+++ epcis-legacy-client/src/main/java/org/oliot/epcis_client/ObjectEvent.java
@@ -27,20 +27,11 @@ import org.bson.BsonDocument;
 public class ObjectEvent extends EPCISEvent {
 
 	// EventTime, EventTimeZoneOffset,Action required
-
 	private List<String> epcList;
 	private List<QuantityElement> quantityList;
 	private String action;
-	private String bizStep;
-	private String disposition;
-	private String readPoint;
-	private String bizLocation;
 	private Map<String, List<String>> bizTransactionList;
-	private Map<String, List<String>> sourceList;
-	private Map<String, List<String>> destinationList;
-	private Map<String, String> namespaces;
 	private Map<String, Map<String, Object>> ilmds;
-	private Map<String, Map<String, Object>> extensions;
 
 	public ObjectEvent() {
 		super();
@@ -49,11 +40,7 @@ public class ObjectEvent extends EPCISEvent {
 		epcList = new ArrayList<String>();
 		quantityList = new ArrayList<QuantityElement>();
 		bizTransactionList = new HashMap<String, List<String>>();
-		sourceList = new HashMap<String, List<String>>();
-		destinationList = new HashMap<String, List<String>>();
-		namespaces = new HashMap<String, String>();
 		ilmds = new HashMap<String, Map<String, Object>>();
-		extensions = new HashMap<String, Map<String, Object>>();
 	}
 
 	public ObjectEvent(long eventTime, String eventTimeZoneOffset, String action) {
@@ -63,11 +50,7 @@ public class ObjectEvent extends EPCISEvent {
 		epcList = new ArrayList<String>();
 		quantityList = new ArrayList<QuantityElement>();
 		bizTransactionList = new HashMap<String, List<String>>();
-		sourceList = new HashMap<String, List<String>>();
-		destinationList = new HashMap<String, List<String>>();
-		namespaces = new HashMap<String, String>();
 		ilmds = new HashMap<String, Map<String, Object>>();
-		extensions = new HashMap<String, Map<String, Object>>();
 	}
 
 	public List<String> getEpcList() {
@@ -94,38 +77,6 @@ public class ObjectEvent extends EPCISEvent {
 		this.action = action;
 	}
 
-	public String getBizStep() {
-		return bizStep;
-	}
-
-	public void setBizStep(String bizStep) {
-		this.bizStep = bizStep;
-	}
-
-	public String getDisposition() {
-		return disposition;
-	}
-
-	public void setDisposition(String disposition) {
-		this.disposition = disposition;
-	}
-
-	public String getReadPoint() {
-		return readPoint;
-	}
-
-	public void setReadPoint(String readPoint) {
-		this.readPoint = readPoint;
-	}
-
-	public String getBizLocation() {
-		return bizLocation;
-	}
-
-	public void setBizLocation(String bizLocation) {
-		this.bizLocation = bizLocation;
-	}
-
 	public Map<String, List<String>> getBizTransactionList() {
 		return bizTransactionList;
 	}
@@ -134,44 +85,12 @@ public class ObjectEvent extends EPCISEvent {
 		this.bizTransactionList = bizTransactionList;
 	}
 
-	public Map<String, List<String>> getSourceList() {
-		return sourceList;
-	}
-
-	public void setSourceList(Map<String, List<String>> sourceList) {
-		this.sourceList = sourceList;
-	}
-
-	public Map<String, List<String>> getDestinationList() {
-		return destinationList;
-	}
-
-	public void setDestinationList(Map<String, List<String>> destinationList) {
-		this.destinationList = destinationList;
-	}
-
-	public Map<String, String> getNamespaces() {
-		return namespaces;
-	}
-
-	public void setNamespaces(Map<String, String> namespaces) {
-		this.namespaces = namespaces;
-	}
-
 	public Map<String, Map<String, Object>> getIlmd() {
 		return ilmds;
 	}
 
 	public void setIlmd(Map<String, Map<String, Object>> ilmds) {
 		this.ilmds = ilmds;
-	}
-
-	public Map<String, Map<String, Object>> getExtensions() {
-		return extensions;
-	}
-
-	public void setExtensions(Map<String, Map<String, Object>> extensions) {
-		this.extensions = extensions;
 	}
 
 	public BsonDocument asBsonDocument() {
@@ -185,37 +104,16 @@ public class ObjectEvent extends EPCISEvent {
 		if (this.epcList != null && this.epcList.size() != 0) {
 			objectEvent = util.putEPCList(objectEvent, epcList);
 		}
-		if (this.bizStep != null) {
-			objectEvent = util.putBizStep(objectEvent, bizStep);
-		}
-		if (this.disposition != null) {
-			objectEvent = util.putDisposition(objectEvent, disposition);
-		}
-		if (this.readPoint != null) {
-			objectEvent = util.putReadPoint(objectEvent, readPoint);
-		}
-		if (this.bizLocation != null) {
-			objectEvent = util.putBizLocation(objectEvent, bizLocation);
-		}
 		if (this.bizTransactionList != null && this.bizTransactionList.isEmpty() == false) {
 			objectEvent = util.putBizTransactionList(objectEvent, bizTransactionList);
 		}
 		if (this.ilmds != null && this.ilmds.isEmpty() == false) {
-			objectEvent = util.putILMD(objectEvent, namespaces, ilmds);
-		}
-		if (this.extensions != null && this.extensions.isEmpty() == false) {
-			objectEvent = util.putExtensions(objectEvent, namespaces, extensions);
+			objectEvent = util.putILMD(objectEvent, getNamespaces(), ilmds);
 		}
 
-		BsonDocument extension = new BsonDocument();
+		BsonDocument extension = asExtensionBsonDocument(util);
 		if (this.quantityList != null && this.quantityList.isEmpty() == false) {
 			extension = util.putQuantityList(extension, quantityList);
-		}
-		if (this.sourceList != null && this.sourceList.isEmpty() == false) {
-			extension = util.putSourceList(extension, sourceList);
-		}
-		if (this.destinationList != null && this.destinationList.isEmpty() == false) {
-			extension = util.putDestinationList(extension, destinationList);
 		}
 		if (extension.isEmpty() == false)
 			objectEvent.put("extension", extension);
