+++ epcis-legacy-client/src/main/java/org/oliot/epcis_client/TransactionEvent.java
@@ -1,7 +1,6 @@
 package org.oliot.epcis_client;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -29,21 +28,9 @@ public class TransactionEvent extends EPCISEvent {
 	// EventTime, EventTimeZoneOffset,Action, bizTransactionList required
 	private String action;
 	private Map<String, List<String>> bizTransactionList;
-
 	private String parentID;
 	private List<String> epcList;
 	private List<QuantityElement> quantityList;
-
-	private String bizStep;
-	private String disposition;
-	private String readPoint;
-	private String bizLocation;
-
-	private Map<String, List<String>> sourceList;
-	private Map<String, List<String>> destinationList;
-	private Map<String, String> namespaces;
-
-	private Map<String, Map<String, Object>> extensions;
 
 	public TransactionEvent(Map<String, List<String>> bizTransactionList) {
 		super();
@@ -54,11 +41,6 @@ public class TransactionEvent extends EPCISEvent {
 
 		epcList = new ArrayList<String>();
 		quantityList = new ArrayList<QuantityElement>();
-
-		sourceList = new HashMap<String, List<String>>();
-		destinationList = new HashMap<String, List<String>>();
-		namespaces = new HashMap<String, String>();
-		extensions = new HashMap<String, Map<String, Object>>();
 	}
 
 	public TransactionEvent(long eventTime, String eventTimeZoneOffset, String action,
@@ -70,11 +52,6 @@ public class TransactionEvent extends EPCISEvent {
 
 		epcList = new ArrayList<String>();
 		quantityList = new ArrayList<QuantityElement>();
-
-		sourceList = new HashMap<String, List<String>>();
-		destinationList = new HashMap<String, List<String>>();
-		namespaces = new HashMap<String, String>();
-		extensions = new HashMap<String, Map<String, Object>>();
 	}
 
 	public String getAction() {
@@ -117,70 +94,6 @@ public class TransactionEvent extends EPCISEvent {
 		this.quantityList = quantityList;
 	}
 
-	public String getBizStep() {
-		return bizStep;
-	}
-
-	public void setBizStep(String bizStep) {
-		this.bizStep = bizStep;
-	}
-
-	public String getDisposition() {
-		return disposition;
-	}
-
-	public void setDisposition(String disposition) {
-		this.disposition = disposition;
-	}
-
-	public String getReadPoint() {
-		return readPoint;
-	}
-
-	public void setReadPoint(String readPoint) {
-		this.readPoint = readPoint;
-	}
-
-	public String getBizLocation() {
-		return bizLocation;
-	}
-
-	public void setBizLocation(String bizLocation) {
-		this.bizLocation = bizLocation;
-	}
-
-	public Map<String, List<String>> getSourceList() {
-		return sourceList;
-	}
-
-	public void setSourceList(Map<String, List<String>> sourceList) {
-		this.sourceList = sourceList;
-	}
-
-	public Map<String, List<String>> getDestinationList() {
-		return destinationList;
-	}
-
-	public void setDestinationList(Map<String, List<String>> destinationList) {
-		this.destinationList = destinationList;
-	}
-
-	public Map<String, String> getNamespaces() {
-		return namespaces;
-	}
-
-	public void setNamespaces(Map<String, String> namespaces) {
-		this.namespaces = namespaces;
-	}
-
-	public Map<String, Map<String, Object>> getExtensions() {
-		return extensions;
-	}
-
-	public void setExtensions(Map<String, Map<String, Object>> extensions) {
-		this.extensions = extensions;
-	}
-
 	public BsonDocument asBsonDocument() {
 		CaptureUtil util = new CaptureUtil();
 
@@ -196,32 +109,10 @@ public class TransactionEvent extends EPCISEvent {
 		if (this.epcList != null && this.epcList.size() != 0) {
 			transactionEvent = util.putEPCList(transactionEvent, epcList);
 		}
-		if (this.bizStep != null) {
-			transactionEvent = util.putBizStep(transactionEvent, bizStep);
-		}
-		if (this.disposition != null) {
-			transactionEvent = util.putDisposition(transactionEvent, disposition);
-		}
-		if (this.readPoint != null) {
-			transactionEvent = util.putReadPoint(transactionEvent, readPoint);
-		}
-		if (this.bizLocation != null) {
-			transactionEvent = util.putBizLocation(transactionEvent, bizLocation);
-		}
 
-		if (this.extensions != null && this.extensions.isEmpty() == false) {
-			transactionEvent = util.putExtensions(transactionEvent, namespaces, extensions);
-		}
-
-		BsonDocument extension = new BsonDocument();
+		BsonDocument extension = asExtensionBsonDocument(util);
 		if (this.quantityList != null && this.quantityList.isEmpty() == false) {
 			extension = util.putQuantityList(extension, quantityList);
-		}
-		if (this.sourceList != null && this.sourceList.isEmpty() == false) {
-			extension = util.putSourceList(extension, sourceList);
-		}
-		if (this.destinationList != null && this.destinationList.isEmpty() == false) {
-			extension = util.putDestinationList(extension, destinationList);
 		}
 		if (extension.isEmpty() == false)
 			transactionEvent.put("extension", extension);
