+++ epcis-legacy-client/src/main/java/org/oliot/epcis_client/TransformationEvent.java
@@ -31,21 +31,9 @@ public class TransformationEvent extends EPCISEvent {
 	private List<QuantityElement> inputQuantityList;
 	private List<String> outputEPCList;
 	private List<QuantityElement> outputQuantityList;
-
 	private String transformationID;
-
-	private String bizStep;
-	private String disposition;
-	private String readPoint;
-	private String bizLocation;
-
 	private Map<String, List<String>> bizTransactionList;
-	private Map<String, List<String>> sourceList;
-	private Map<String, List<String>> destinationList;
-
-	private Map<String, String> namespaces;
 	private Map<String, Map<String, Object>> ilmds;
-	private Map<String, Map<String, Object>> extensions;
 
 	public TransformationEvent() {
 		super();
@@ -56,11 +44,7 @@ public class TransformationEvent extends EPCISEvent {
 		outputQuantityList = new ArrayList<QuantityElement>();
 
 		bizTransactionList = new HashMap<String, List<String>>();
-		sourceList = new HashMap<String, List<String>>();
-		destinationList = new HashMap<String, List<String>>();
-		namespaces = new HashMap<String, String>();
 		ilmds = new HashMap<String, Map<String, Object>>();
-		extensions = new HashMap<String, Map<String, Object>>();
 	}
 
 	public TransformationEvent(long eventTime, String eventTimeZoneOffset) {
@@ -72,11 +56,7 @@ public class TransformationEvent extends EPCISEvent {
 		outputQuantityList = new ArrayList<QuantityElement>();
 
 		bizTransactionList = new HashMap<String, List<String>>();
-		sourceList = new HashMap<String, List<String>>();
-		destinationList = new HashMap<String, List<String>>();
-		namespaces = new HashMap<String, String>();
 		ilmds = new HashMap<String, Map<String, Object>>();
-		extensions = new HashMap<String, Map<String, Object>>();
 	}
 
 	public List<String> getInputEPCList() {
@@ -119,38 +99,6 @@ public class TransformationEvent extends EPCISEvent {
 		this.transformationID = transformationID;
 	}
 
-	public String getBizStep() {
-		return bizStep;
-	}
-
-	public void setBizStep(String bizStep) {
-		this.bizStep = bizStep;
-	}
-
-	public String getDisposition() {
-		return disposition;
-	}
-
-	public void setDisposition(String disposition) {
-		this.disposition = disposition;
-	}
-
-	public String getReadPoint() {
-		return readPoint;
-	}
-
-	public void setReadPoint(String readPoint) {
-		this.readPoint = readPoint;
-	}
-
-	public String getBizLocation() {
-		return bizLocation;
-	}
-
-	public void setBizLocation(String bizLocation) {
-		this.bizLocation = bizLocation;
-	}
-
 	public Map<String, List<String>> getBizTransactionList() {
 		return bizTransactionList;
 	}
@@ -159,44 +107,12 @@ public class TransformationEvent extends EPCISEvent {
 		this.bizTransactionList = bizTransactionList;
 	}
 
-	public Map<String, List<String>> getSourceList() {
-		return sourceList;
-	}
-
-	public void setSourceList(Map<String, List<String>> sourceList) {
-		this.sourceList = sourceList;
-	}
-
-	public Map<String, List<String>> getDestinationList() {
-		return destinationList;
-	}
-
-	public void setDestinationList(Map<String, List<String>> destinationList) {
-		this.destinationList = destinationList;
-	}
-
-	public Map<String, String> getNamespaces() {
-		return namespaces;
-	}
-
-	public void setNamespaces(Map<String, String> namespaces) {
-		this.namespaces = namespaces;
-	}
-
 	public Map<String, Map<String, Object>> getIlmds() {
 		return ilmds;
 	}
 
 	public void setIlmds(Map<String, Map<String, Object>> ilmds) {
 		this.ilmds = ilmds;
-	}
-
-	public Map<String, Map<String, Object>> getExtensions() {
-		return extensions;
-	}
-
-	public void setExtensions(Map<String, Map<String, Object>> extensions) {
-		this.extensions = extensions;
 	}
 
 	public BsonDocument asBsonDocument() {
@@ -222,35 +138,14 @@ public class TransformationEvent extends EPCISEvent {
 			transformationEvent = util.putTransformationID(transformationEvent, transformationID);
 		}
 
-		if (this.bizStep != null) {
-			transformationEvent = util.putBizStep(transformationEvent, bizStep);
-		}
-		if (this.disposition != null) {
-			transformationEvent = util.putDisposition(transformationEvent, disposition);
-		}
-		if (this.readPoint != null) {
-			transformationEvent = util.putReadPoint(transformationEvent, readPoint);
-		}
-		if (this.bizLocation != null) {
-			transformationEvent = util.putBizLocation(transformationEvent, bizLocation);
-		}
 		if (this.bizTransactionList != null && this.bizTransactionList.isEmpty() == false) {
 			transformationEvent = util.putBizTransactionList(transformationEvent, bizTransactionList);
 		}
 		if (this.ilmds != null && this.ilmds.isEmpty() == false) {
-			transformationEvent = util.putILMD(transformationEvent, namespaces, ilmds);
-		}
-		if (this.extensions != null && this.extensions.isEmpty() == false) {
-			transformationEvent = util.putExtensions(transformationEvent, namespaces, extensions);
+			transformationEvent = util.putILMD(transformationEvent, getNamespaces(), ilmds);
 		}
 
-		BsonDocument extension = new BsonDocument();
-		if (this.sourceList != null && this.sourceList.isEmpty() == false) {
-			extension = util.putSourceList(extension, sourceList);
-		}
-		if (this.destinationList != null && this.destinationList.isEmpty() == false) {
-			extension = util.putDestinationList(extension, destinationList);
-		}
+		BsonDocument extension = asExtensionBsonDocument(util);
 		if (extension.isEmpty() == false)
 			transformationEvent.put("extension", extension);
